fix: coin register functions
refactor: coin regististration

fix: unregister only meta coins

fix: move register coin into add_pool*

fix: remove pool from state tracker

fix: update location of coin_b/c when rm coins
